Major architecture changes. Some fixes and improvements.
